Create README.md
Assignment 2: 50 pts
You now work for Titan Dev. We are a software development company that makes business software for any type of corporation. We have been asked to build a payroll system for Walmart. The following are some notes we took while conversing with our customer. 
This system consists of a database of the company’s employees, and their associated data, such as time cards. The system must pay all employees the correct amount, on time by the method that they specify. Also, various deductions must be taken from their pay.
•	Some employees work by the hour. They are paid an hourly rate that is one of the fields in their employee record. They submit daily time cards that record the date and the number of hours worked. If they work more than 8 hours per day, they are paid 1.5 times their normal rate for those extra hours. They are paid every Friday.
•	Some employees are paid a flat salary. They are paid on the last working day of the month. Their salary is one of the fields in their employee record.
•	Some of the salaried employees are also paid a commission based on their sales. They submit receipts that record the date and the amount of the sale. Their commission rate is a field in their employee record. They are paid every other Friday.
•	Employees can select their method of payment. They may have their paychecks mailed to the postal address of their choice, have their paychecks held by the paymaster for pickup, or request that their paychecks be directly deposited into the bank account of their choice.
•	Some employees belong to the union. Their employee record has a field for the weekly dues rate. Their dues must be deducted from their pay. Also, the union may assess service charges against individual union members from time to time. These service charges are submitted by the union on a weekly basis and must be deducted from the appropriate employee’s next pay amount.
•	The payroll application will run once each working day and pay the appropriate employees on that day. The system will be told what date the employees are to be paid to, so it will generate payments for records from the last time the employee was paid up to the specified date.
Prerequisites:
To complete this assignment, you should have pycharm installed and have created a Github account already.
Task:
Gone are the days when we worked as an individual in the software development world. Now you must be able to work effectively in a team environment. This means you need to be able to use tools such as source control effectively as well. For this assignment, we will use Github as our source control repository. First you need to find a partner. Post a partner request or respond to a partner request in the partner forum in module 2. Once you’ve found a partner, decide together who will create the initial github repository. Add your partner as a collaborator to the repository. The development requirements for this assignment are below:
You have identified the initial classes in the UML diagram below. Create an application called “TitanPay”.
Create a module called “accounting” in your project. In the “accounting” module, create the following classes and implement their attributes and behavior.
 
Notes: 
The get_full_name method on the Employee and SalariedEmployee classes should return the first and lastname in the following format: example: Tillman, Adrian
The get_address method on the Address class should return the entire address as follows:
Example: 6006 5 Ave. N. St. Petersburg, Fl. 33712
The calculate_daily_pay(rate) on the TimeCard, takes the normal rate of pay and calculates the pay for the day according to the following rule. If they work more than 8 hours per day, they are paid 1.5 times their normal rate for those extra hours. If 8 or less, they receive their normal pay rate times the hours worked.
The “deposit(amt)” method on the BankAccount, takes the amt and will print to the screen the following message. Example: Depositing $x.xx in Bank of America Account Number: xxxxxxxx using Routing Number: xxxxxxxx
Split the work equally among you and your partner. Be sure to continuously commit and push your code to your repository so each partner has the current version of the code. When complete, each partner should post a word document to the dropbox that contains what they worked on in the project and the github repository link. Each partner is responsible for completing three of the classes outlined above.  
If you don’t know how to push to your GitHub repository, read the link in the supplements folder or watch the example lecture, “Submitting to Github”.